bigInterger > BitInteger

+++ sampleapp/myapp/apps/main/models/tables/usertables.py
@@ -3,7 +3,7 @@ from .use_sqlalchemy import *
 memo = Table(
     "memo", db.metadata,
     Column('id', BigInteger, primary_key=True),
-    Column('user_id', bigInteger, ForeignKey('user.id')),
+    Column('user_id', BigInteger, ForeignKey('user.id')),
     Column('text', UnicodeText(), default=u"", nullable=False,),
     Column('deleted', Boolean(), default=False),
     )
